requirements.txt
requests
pandas
scikit-learn
python-dotenv
google-generativeai
fuzzywuzzy
python-Levenshtein
